Close mongo client before return
(After installing mongo 😅) I was able to repro Eleventy hanging locally and figured the only "weird" bit was the `async function getFilms`, so I started poking around there.

Commenting it out the whole file allowed 11ty to return properly, so I was pretty sure this was it. I saw that the MongoDB `client` was opened, but never closed, which might explain why 11ty never exited. So I tried closing the client after the data resolves and it seems to have worked.